DigitalClock
Getting Set Up: We start by getting everything ready, like making sure we have what we need to tell the time.

Running Forever: Then, we start doing something over and over again, like a never-ending loop.

Updating the Time: Inside this loop, we keep track of the time. We count seconds, and whenever we get to 60, we start over at 0 and increase the minutes. We do the same for minutes and hours, making sure they don't go over their limits.

Showing the Time: After we know what time it is, we show it, like writing it down on a piece of paper.

Doing it Again: We then wait for a little bit (1 second), and then start over again from the beginning, so we can keep showing the correct time.